requirements.txt
langchain==0.0.340
bs4==0.0.1
openai==1.3.5
chromadb==0.4.18
langchainhub==0.1.14
tiktoken==0.5.1
pyyaml==6.0.1
fastapi==0.104.1
datasets==2.15.0
sentence-transformers==2.2.2
InstructorEmbedding==1.0.1